some early drafts of debuerreo'ing.
It's got an *extra* PITA because the current bootstrap asset has no wget, which is just infuriating.  We'll escape this when we get this thing to chase its tail.

This doesn't reproduce yet, either.  But at least it does repeat.

Oh, and the output size is 100MB gzipped.  Which is pretty awful, and I dearly hope we can improve upon radically.  It should be less than a quarter of that, in my figuring.

Signed-off-by: Eric Myhre <hash@exultant.us>