it would be nice if we could split the mknod part.
Putting it in a separate formula would let us run the debootstrap parts -- which contact a network, natch -- in a lower-priv mode, and that would be nice.

Of course, this doesn't work.  The debootstrap script checks whether it can mknod, regardless of whether or not it needs it; and also it simply refuses to operate on an existing directory, so we can't volunteer to do the work for it.

Signed-off-by: Eric Myhre <hash@exultant.us>