we can have yaml now!
multi-line strings really are necessary to not lose your mind after a while.

We could also have just put the script in another file nearby, and used a scan or similar technique to get it into place, but that's less fun, and it's often nice to *see* the whole script in one go.

Signed-off-by: Eric Myhre <hash@exultant.us>